fix(rust): unblock CI by repairing long-standing build failures
The astroforge-core crate has never compiled since the project was
initialised. CI has been red on every push and PR since
a9d6b6f "Phase 0 M1: Initialize Tauri 2.x project skeleton".

Root cause: src/image.rs wrote `impl F32Image { ... }` for
`pub type F32Image = Array3<f32>`, violating Rust's orphan rule.
All 20+ files calling `.width()`, `.height()`, `.channels()`,
`F32Image::new()` failed to resolve, blocking 255+ errors.

The fmt error in star_segmentation.rs:85 (unclosed paren) was just
the surface symptom — once cargo fmt could read the file, it
uncovered a much deeper build failure that had always been there.

Fix scope:

* crates/astroforge-core/src/image.rs
  - Rewrite `F32Image` as a newtype wrapper around `Array3<f32>`
    so we can define inherent methods without orphan-rule violation
  - Add `Deref`/`DerefMut` so `img[(c, y, x)]` indexing and all
    other `Array3` methods keep working unchanged at call sites
  - Forward `Sub` (for `dark - bias` pixel-wise arithmetic),
    `Serialize`, `Deserialize` (for the `StackResult` struct)
  - Add `From<Array3<f32>>` and back for interop

* crates/astroforge-core/Cargo.toml
  - Enable ndarray `serde` feature so the manual `Serialize`/
    `Deserialize` impls above have something to delegate to

* Cargo.lock
  - Pick up the new `ndarray` feature resolution

* crates/astroforge-core/src/export.rs
  - Drop `as u16` cast on `channels` (passed into a `usize` param)
  - Pre-existing bug — never compiled

* crates/astroforge-core/src/stretching.rs
  - Cast `normalized: f32` to `f64` before passing to
    `arcsinh_stretch(value: f64, ...)`
  - Pre-existing bug — never compiled

* Cargo fmt drift across 32 files (mechanical use-list reorder,
  signature reflow) from running rustfmt 1.98 (CI's toolchain)
  over a tree last formatted with 1.81

* Pre-existing clippy lints cleared (13 sites):
  - Bayer detection identical if-branches (real bug: 0.5–0.85
    case collapsed into the >0.85 case)
  - debayer RGGB/BGGR duplicate green branches (collapsed via XOR)
  - planetary_routing dead `&& count < 500` condition
  - fits unused `use std::io::Cursor` import in test mod
  - calibration unused `mut frames`
  - quality `.max(0.0).min(1.0)` -> `.clamp(0.0, 1.0)`
  - recipe type-complexity: `Vec<(String, HashMap<...>)>` extracted
    as `pub type StageParams`
  - planetary_lucky `.len() >= 1` -> `!is_empty()`
  - debayer `from_str` -> `parse` (avoid `FromStr` confusion);
    updated 3 call sites in dng_parser and the test
  - astroforge-ai models `.score >= 0.0 && score <= 1.0` ->
    `(0.0..=1.0).contains(&score)`
  - color_calibration / debayer `for c in 0..n { gains[c] }`
    -> `for (c, &gain) in gains.iter().take(n).enumerate()`
  - tiling test bug: `weight[(0, 0, 0)] > 0.0` checked at a
    corner where the cosine blend weight is exactly 0;
    moved to `(0, 4, 4)` where the window weight is positive
  - astroforge-core mvp_pipeline collapsed two adjacent
    `#[derive(...)]` lines into one (rustfmt)

* `cargo fix --workspace` cleanup of unused imports / unused
  variables (the easy wins).

What remains:
- 13 implementation-level test failures (e.g. `result.blue_gain
  < 1.0`, `denoised within 0.5 of original`, histogram bin counts
  zero where 16 expected, stacking rejecting an outlier returning
  0 where 100 expected). These are pre-existing logic bugs that
  have never been caught because the crate never compiled. They
  are out of scope for this PR.

Verified locally with the same toolchain CI uses:
  cargo +stable fmt --all -- --check       # exit 0
  cargo +stable clippy --workspace --all-targets -- -D warnings   # exit 0
  cargo +stable check --workspace          # exit 0

Unblocks:
- #188 (chore/phase1.5-plan-crosslinks)
- #189 (docs/design: propose AIService interface for M1.5.5 T1)

Once CI is green on this PR and it merges, both should merge
cleanly.

+++ crates/astroforge-core/src/stretching.rs
@@ -11,7 +11,7 @@ pub fn auto_stretch(image: &F32Image) -> F32Image {
 
     for val in result.iter_mut() {
         let normalized = (*val - min) / range;
-        *val = arcsinh_stretch(normalized, midtones) as f32;
+        *val = arcsinh_stretch(f64::from(normalized), midtones) as f32;
     }
 
     result
